requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.4.2
Flask==1.0.2
future==0.17.1
googletrans==2.4.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
line-bot-sdk==1.8.0
MarkupSafe==1.1.0
pycparser==2.19
pymongo==3.7.2
pyOpenSSL==18.0.0
requests==2.21.0
six==1.12.0
SpeechRecognition==3.8.1
urllib3==1.24.1
Werkzeug==0.14.1
